created util package to contain all the util classes
