--- runner/src/usecase/BUILD
@@ -1,3 +1,0 @@
-java_library(
-    name = "usecase"
-)
