+++ runner/src/util/BenchmarkConfigurationFile.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.benchmark.runner.executionconfig;
+package grakn.benchmark.runner.util;
 
 
 import java.util.List;
